+++ app/src/main/java/com/example/appquanlychitieu/data/model/GoalHistory.java
@@ -4,12 +4,12 @@ public class GoalHistory {
     private long id;
     
     private long goalId;
-    private double amountAdded;
+    private long amountAdded;
     private long date;
     private String remoteId;
     private String remoteGoalId;
 
-    public GoalHistory(long goalId, double amountAdded, long date) {
+    public GoalHistory(long goalId, long amountAdded, long date) {
         this.goalId = goalId;
         this.amountAdded = amountAdded;
         this.date = date;
@@ -21,8 +21,8 @@ public class GoalHistory {
     public long getGoalId() { return goalId; }
     public void setGoalId(long goalId) { this.goalId = goalId; }
 
-    public double getAmountAdded() { return amountAdded; }
-    public void setAmountAdded(double amountAdded) { this.amountAdded = amountAdded; }
+    public long getAmountAdded() { return amountAdded; }
+    public void setAmountAdded(long amountAdded) { this.amountAdded = amountAdded; }
 
     public long getDate() { return date; }
     public void setDate(long date) { this.date = date; }
