+++ app/src/main/java/com/example/appquanlychitieu/data/remote/dto/BudgetRequestDto.java
@@ -1,13 +1,12 @@
 package com.example.appquanlychitieu.data.remote.dto;
 
-import java.math.BigDecimal;
 
 public class BudgetRequestDto {
     public String categoryId;
-    public BigDecimal amount;
+    public long amount;
     public String monthYear;
 
-    public BudgetRequestDto(String categoryId, BigDecimal amount, String monthYear) {
+    public BudgetRequestDto(String categoryId, long amount, String monthYear) {
         this.categoryId = categoryId;
         this.amount = amount;
         this.monthYear = monthYear;
