+++ app/src/main/java/com/example/appquanlychitieu/data/remote/dto/CategoryStatisticDto.java
@@ -1,12 +1,11 @@
 package com.example.appquanlychitieu.data.remote.dto;
 
-import java.math.BigDecimal;
 
 public class CategoryStatisticDto {
     public String categoryId;
     public String categoryName;
     public String type;
-    public BigDecimal total;
+    public Long total;
     public int transactionCount;
     public String categoryColor;
     public String categoryIcon;
