+++ app/src/main/java/com/example/appquanlychitieu/data/remote/dto/GoalHistoryDto.java
@@ -1,10 +1,9 @@
 package com.example.appquanlychitieu.data.remote.dto;
 
-import java.math.BigDecimal;
 
 public class GoalHistoryDto {
     public String id;
     public String goalId;
-    public BigDecimal amountAdded;
+    public Long amountAdded;
     public String date;
 }
