+++ app/src/main/java/com/example/appquanlychitieu/data/remote/dto/GoalRequestDto.java
@@ -1,13 +1,12 @@
 package com.example.appquanlychitieu.data.remote.dto;
 
-import java.math.BigDecimal;
 
 public class GoalRequestDto {
     public String name;
-    public BigDecimal targetAmount;
-    public BigDecimal currentAmount;
+    public long targetAmount;
+    public long currentAmount;
 
-    public GoalRequestDto(String name, BigDecimal targetAmount, BigDecimal currentAmount) {
+    public GoalRequestDto(String name, long targetAmount, long currentAmount) {
         this.name = name;
         this.targetAmount = targetAmount;
         this.currentAmount = currentAmount;
