+++ app/src/main/java/com/example/appquanlychitieu/data/remote/dto/MonthlyStatisticDto.java
@@ -1,11 +1,10 @@
 package com.example.appquanlychitieu.data.remote.dto;
 
-import java.math.BigDecimal;
 
 public class MonthlyStatisticDto {
     public int year;
     public int month;
-    public BigDecimal income;
-    public BigDecimal expense;
-    public BigDecimal balance;
+    public Long income;
+    public Long expense;
+    public Long balance;
 }
